requirements/common.txt
asgiref==3.8.1
asttokens==2.4.1
boto3==1.35.10
botocore==1.35.20
cffi==1.17.1
click==8.1.7
coverage==7.6.1
cryptography==43.0.1
decorator==5.1.1
dj-database-url==2.2.0
Django==5.1.1
django-configurations==2.5.1
django-filter==24.3
django-model-utils==4.5.1
django-storages==1.14.4
django-unique-upload==0.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
factory_boy==3.3.1
Faker==28.4.1
flake8==7.1.1
ghp-import==2.1.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mock==5.1.0
newrelic==9.13.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.3
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==5.0.2
wcwidth==0.2.13
typing==3.7.4.3
